move common code to types/utils
