+++ src/scripts/tbtcv2-rewards/types.ts
@@ -1,0 +1,5 @@
+export type InstanceParams = {
+  upTimePercent: number
+  avgPreParams: number
+  version: any
+}
